lumina_ai/requirements.txt
django>=5.0
djangorestframework
djangorestframework-simplejwt
python-dotenv
requests
django-cors-headers
